+++ cypress/support/newContractMemberCommands.js
@@ -1,3 +1,10 @@
+Cypress.Commands.add('identifyMember', ({ nif, number }) => {
+  cy.get('[data-cy="vat"]').type(nif)
+  cy.get('[data-cy="code"]').type(number)
+
+  cy.get('[data-cy=next]').click()
+})
+
 Cypress.Commands.add('contractMemberQuestion', (has_member = false) => {
   const optionValue = has_member ? 'member-on' : 'member-off'
   cy.get('[data-cy="member-question"]')
@@ -5,13 +12,22 @@ Cypress.Commands.add('contractMemberQuestion', (has_member = false) => {
     .click()
 })
 
-Cypress.Commands.add('contractMemberHolderQuestion', (previous_holder = true) => {
-  const optionValue = previous_holder ? 'previous-holder-yes' : 'previous-holder-no'
-  cy.get('[data-cy="holder-question"]')
-    .get(`[data-cy="${optionValue}"]`)
-    .click()
-  
-  cy.get('[data-cy=next]').click()
+Cypress.Commands.add('contractMemberHolderQuestion', (has_light = "light-on", has_member = true, previous_holder = true, member_is_holder = true) => {
+  if (has_light === "light-on") {
+    const holderQuestionValue = previous_holder ? 'previous-holder-yes' : 'previous-holder-no'
+    cy.get('[data-cy="holder-question"]')
+      .get(`[data-cy="${holderQuestionValue}"]`)
+      .click()
+  }
+  if (!has_member){
+    const holderMemberQuestionValue = member_is_holder ? 'holder-member-yes' : 'holder-member-no'
+    cy.get('[data-cy="holder-member-question"]')
+      .get(`[data-cy="${holderMemberQuestionValue}"]`)
+      .click()
+  }
+  if (has_light === "light-on" || !has_member){
+    cy.get('[data-cy=next]').click()
+  }
 })
 
 Cypress.Commands.add('contractMemberDonationQuestion', (donation = true) => {
@@ -40,6 +56,10 @@ Cypress.Commands.add('newContractIdentifySupplyPoint', (cups, has_light, statusC
   cy.get('[data-cy="cups"]').type(cups)
 
   cy.get('[data-cy="light-question"]').get(`[data-cy="${has_light}"]`).click()
+
+  if (has_light === "light-off"){
+    cy.get('[data-cy=accept]').click()
+  }
 
   cy.get('[data-cy=next]').click()
 })
@@ -97,10 +117,12 @@ Cypress.Commands.add('selfconsumptionData', (selfConsumption) => {
 
 })
 
-Cypress.Commands.add('contractMemberPaymentData', (paymentdata) => {
+Cypress.Commands.add('contractMemberPaymentData', (paymentdata, has_member=false) => {
   cy.get('[data-cy="iban_number"]').type(paymentdata.iban)
 
-  cy.choosePaymentMethod()
+  if (!has_member){
+    cy.choosePaymentMethod()
+  }
 
   cy.get('[data-cy="iban_check"]').click()
   cy.get('[data-cy=accept]').click()
@@ -108,14 +130,16 @@ Cypress.Commands.add('contractMemberPaymentData', (paymentdata) => {
   cy.get('[data-cy=next]').click()
 })
 
-Cypress.Commands.add('contractMemberCheckReviewNewMemberStep', (nif) => {
+Cypress.Commands.add('contractMemberCheckReviewNewMemberStep', (nif, has_member=false) => {
   cy.get('[data-cy="privacy_policy"]').click()
 
   cy.get('[data-cy="generic_conditions_accepted"]').click()
   cy.get('[data-cy="generic_conditions_modal"]').scrollTo('bottom')
   cy.get('[data-cy=accept]').click()
 
-  cy.get('[data-cy="statutes"]').click()
+  if (!has_member){
+    cy.get('[data-cy="statutes"]').click()
+  }
 
   cy.get('[id="edit_button"]').click()